directly reduce ensemble predictions

+++ scripts/test/test_ndvi.py
@@ -192,7 +192,7 @@ def main(
     
     print(f"Creating data for calibration plot...")
     processes = min(mp.cpu_count(), processes) 
-    df_calibration = create_calibration_plot(df, scipy.stats.norm, processes=processes, num_samples=len(df) * 0.5)
+    df_calibration = create_calibration_plot(df, scipy.stats.norm, processes=processes, num_samples=int(len(df) * 0.5))
     df_calibration.to_csv(result_dir / f"calibration.csv", index=False)
 
     print(f"Finished processing dataset!")
